services/listener/src/requirements.txt
redis
discord
python-dotenv
pylint
